Версия 3
- Функция, меняющая любой тип данных между собой местами (быстрая, меняет по 8 байт за раз). Предыдущая функция удалена

- Функция, создающая массив с текстом из файла

- Функция, считающая количество строк в массиве и заменяющая \n на \0

- Функция распечатки массива стала быстрее (теперь реализована через puts, а не использует printf ля посимвольного вывода)

+++ include/Structs.h
@@ -6,7 +6,7 @@
 struct PtrAndLenString
     {
     char* ptr_str; // указатель на строку
-    int len_str; // длина строки
+    size_t len_str; // длина строки
     };
 
 const size_t SIZE_PtrAndLenString = sizeof (PtrAndLenString);
